Restructure NodeJS wrapper files
