+++ wrappers/nodejs/package.json
@@ -43,7 +43,7 @@
         "@typescript-eslint/eslint-plugin": "^3.3.0",
         "@typescript-eslint/parser": "^3.3.0",
         "app-module-path": "^2.2.0",
-        "axios": "^0.20.0",
+        "axios": "^0.19.0",
         "body-parser": "^1.19.0",
         "canonical-json": "0.0.4",
         "chai": "^4.1.2",
@@ -65,6 +65,6 @@
         "test:integration": "TS_NODE_PROJECT=./test/tsconfig.json NODE_ENV=test RUST_BACKTRACE=full mocha --timeout 10000 --expose-gc --exit --recursive --use_strict --require ts-node/register ./test/integration/*.spec.ts",
         "test:ffi-tools": "TS_NODE_PROJECT=./test/tsconfig.json NODE_ENV=test RUST_BACKTRACE=full mocha --timeout 10000 --expose-gc --exit --recursive --use_strict --require ts-node/register ./test/unit/ffi-tools.spec.ts",
         "test:vdr-pool": "TS_NODE_PROJECT=./test/tsconfig.json NODE_ENV=test RUST_BACKTRACE=full mocha --timeout 10000 --expose-gc --exit --recursive --use_strict --require ts-node/register ./test/unit/indy-vdr-pool.spec.ts",
-        "test:vdr-request": "TS_NODE_PROJECT=./test/tsconfig.json NODE_ENV=test RUST_BACKTRACE=full mocha --timeout 10000 --expose-gc --exit --recursive --use_strict --require ts-node/register ./test/unit/indy-vdr-request.spec.ts"
+        "test:vdr-request": "TS_NODE_PROJECT=./test/tsconfig.json NODE_ENV=test RUST_BACKTRACE=full mocha --timeout 10000 --expose-gc --exit --recursive --use_strict --require ts-node/register ./test/unit/ledger-requests/*.spec.ts"
     }
 }
