+++ wrappers/nodejs/src/index.ts
@@ -1,6 +1,6 @@
 export * from './api/indy-vdr-pool';
-export * from './api/indy-vdr-request';
 export * from './api/indy-vdr-utils';
+export * from './api/ledger-requests/ledger-request-custom';
 export * from './vdr';
 export * from './rustlib';
 export * from './errors';
